+++ buildchain/vite.config.ts
@@ -1,9 +1,9 @@
 import {defineConfig} from 'vite';
 import {visualizer} from 'rollup-plugin-visualizer';
-import viteEslintPlugin from 'vite-plugin-eslint';
+import checker from 'vite-plugin-checker';
+import tailwindcss from "@tailwindcss/vite";
 import viteCompressionPlugin from 'vite-plugin-compression';
 import viteRestartPlugin from 'vite-plugin-restart';
-import viteStylelintPlugin from 'vite-plugin-stylelint';
 import * as path from 'path';
 
 // https://vitejs.dev/config/
@@ -36,14 +36,27 @@ export default defineConfig(({command}) => ({
       template: 'treemap',
       sourcemap: true,
     }),
-    viteEslintPlugin({
-      cache: false,
-      fix: true,
+    tailwindcss(),
+    checker({
+      eslint: {
+        lintCommand: 'eslint "./src/**/*.{js,ts}"',
+        useFlatConfig: true,
+        dev: {
+          overrideConfig: {
+            cache: true,
+          }
+        }
+      },
+      stylelint: {
+        lintCommand: 'stylelint ./src/**/*.{css,scss,sass,pcss} --fix',
+        dev: {
+          overrideConfig: {
+            cache: true,
+          }
+        }
+      },
+      typescript: true,
     }),
-    viteStylelintPlugin({
-      fix: true,
-      lintInWorker: true
-    })
   ],
   resolve: {
     alias: [
@@ -52,8 +65,16 @@ export default defineConfig(({command}) => ({
     preserveSymlinks: true,
   },
   server: {
+    // Allow cross-origin requests -- https://github.com/vitejs/vite/security/advisories/GHSA-vg6x-rcgg-rjx6
+    allowedHosts: true,
+    cors: {
+      origin: /https?:\/\/([A-Za-z0-9\-\.]+)?(localhost|\.local|\.test|\.ddev\.site)(?::\d+)?$/
+    },
     fs: {
       strict: false
+    },
+    headers: {
+      "Access-Control-Allow-Private-Network": "true",
     },
     host: '0.0.0.0',
     origin: 'http://localhost:' + process.env.DEV_PORT,