+++ src/models/CantoFieldData.php
@@ -5,12 +5,22 @@ namespace lsst\cantodamassets\models;
 use craft\base\Model;
 use craft\validators\ArrayValidator;
 use lsst\cantodamassets\lib\laravel\Collection;
+use yii\helpers\Inflector;
 
 /**
  * Canto DAM Field Data
  */
 class CantoFieldData extends Model
 {
+    /**
+     * Which fields have their properties camelized to be compatible with GraphQL query params
+     */
+    public const CAMELIZED_FIELDS = [
+        'metadata',
+        'additional',
+        'default',
+    ];
+
     public ?string $cantoId = null;
     public ?string $cantoAlbumId = null;
     public Collection|array $cantoAssetData = [];
@@ -31,6 +41,19 @@ class CantoFieldData extends Model
     public function init(): void
     {
         parent::init();
+        // Make sure we camelize the keys if an array is being returned, since we normalize them to be camelized
+        // as GraphQL doesn't support spaces or other special characters in the query params
+        foreach ($this->cantoAssetData as &$item) {
+            foreach (self::CAMELIZED_FIELDS as $fieldName) {
+                if (isset($item[$fieldName])) {
+                    foreach ($item[$fieldName] as $key => $value) {
+                        unset($item[$fieldName][$key]);
+                        $item[$fieldName][Inflector::camelize($key)] = $value;
+                    }
+                }
+            }
+        }
+
         $this->cantoAssetData = new Collection($this->cantoAssetData);
         $this->cantoAlbumData = new Collection($this->cantoAlbumData);
     }